pyproject.toml
[tool.poetry]
name = "recommendation-system-api"
version = "0.1.0"
description = ""
authors = ["Thiennn262626 <nguyendinhthien10a5@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
Flask = "^3.0.2"
sqlalchemy = "^2.0.29"
flask-restful = "^0.3.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
